requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.30
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.0
Werkzeug>=3.0.0
Flask-Bcrypt>=1.0.1
Flask-Login>=0.6.3
email-validator>=2.0.0
# Database drivers - uncomment the one you need:
# psycopg2-binary==2.9.7  # PostgreSQL
# PyMySQL==1.1.0          # MySQL
# pyodbc==4.0.39          # SQL Server
